673518 - Store consumer credentials on Consumer and use to prevent wrong cleanup. First thought of using certificate serial number so added next_serial() to make sure it was unique.  But then decided to just store the key and cert.  But, already did the work so decided to leave it.  It's proper to have unique SN anyway on x.509. Also, decided to cleanup the /etc/yum.repos.d/pulp.repo when consumer is deleted.

+++ src/pulp/client/gofer/pulp.py
@@ -103,13 +103,21 @@ class Consumer:
     """
     
     @remote
-    def deleted(self):
+    def deleted(self, credentials):
         """
         Notification that the consumer has been deleted.
         Clean up associated artifacts.
+        @param credentials: The associated x.509 credentials.
+        @type credentials: tuple (key,crt)
         """
         bundle = ConsumerBundle()
+        found = bundle.read()
+        expected = tuple(credentials)
+        if found != expected:
+            return
         bundle.delete()
+        repo = Repo()
+        repo.delete()
         log.info('Artifacts deleted')
 
 
@@ -127,6 +135,14 @@ class Repo:
         log.info('updating yum repo')
         rlib = RepoLib()
         rlib.update()
+
+    @remote
+    def delete(self):
+        """
+        Delete the .repo file.
+        """
+        rlib = RepoLib()
+        rlib.delete()
 
 
 class Packages:
